extracted positioning, aka dragging and grabbing, into modules
